Initial pass at web application
Create a simple Backbone app that grabs untranscribed cells and
allows the user to easily input text values.

Update the semantics of the ``/cells`` endpoint so that the
limit parameter is ``text_lt`` instead of ``text_lte``.  That is,
the results will be limited to cells with fewer than the given
number of text transcriptions.