Web views, database-based merging workflow
Add Flask views to be able to view projects, source files,
page images and cells.

Rename the ``merge_ocr`` command to be just ``merge`` and
update the command to pull text from the from the database
instead of the OCR output text files.

Update models with utility methods to support the new views and
the merging workflow.

Update JavaScripts to handle case where there are no more cells
that need transcribing.